test(arraydeque): split benchmarks into dedicated file and add rigorous benchmarking scenarios
